+++ src/App.vue
@@ -1,23 +1,9 @@
 <script lang="ts" setup>
-import { reactive } from "vue";
 
-const state = reactive({ name: "" });
-const startGame = () => {
-  alert(`ユーザ名：${state.name}`);
-}
 </script>
 
 <template>
-  <h1>トップ画面</h1>
-  <p>{{ state.name }}</p>
-  <input
-    v-model="state.name"
-    placeholder="名前を入力してください (10文字以内)"
-    maxlength="10"
-  ><br><br>
-  <button v-on:click="startGame()">
-    ゲームスタート！
-  </button>
+  <router-view />
 </template>
 
 <!-- state.name 確認用のタグのスタイルを指定する -->
